# robert-client-api-spec/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "#STOPCOVID, Robert Proximity Tracing Client API",
        "version": "1.0.0",
        "title": "#STOPCOVID, Robert Proximity Tracing",
        "termsOfService": "https://github.com/ROBERT-proximity-tracing/",
        "contact": {
            "email": "stopcovid@inria.fr"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "yourrobertserver.org",
    "basePath": "/v1",
...
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/unregister": {
            "post": {
                "tags": [
                ],
                "summary": "Unregister app",
                "description": "User wants to stop using the ROBERT service. Delete all data related to this user from the server.\n\nThis request must use the '0x03' salt when creating its MAC digest",
                "operationId": "unregister",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "req",
                        "schema":{
                            "$ref": "#/definitions/AuthentifiedRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns whether unregistering was successful",
                        "schema": {
                            "$ref": "#/definitions/UnregisterResponse"
                        }
                    },
                    "400": {
                        "description": "bad request"
                    },
                    "401": {
                        "description": "invalid authentication"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
    },
    "definitions": {
        "UnregisterResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            },
            "required": [
                "success"
            ]
        },
        "AuthentifiedRequest": {
            "type": "object",
            "properties": {
                "ebid": {
                    "description": "Any EBID from the local store valid for this epoch as 64-bit, Base64 encoded\n\nRaw content must be 64 bits. Base64 string must be 12 chars.",
                    "type": "string",
                    "format": "byte"
                },
                "time": {
                    "description": "The current time as 32-bit timestamp, Base64 encoded\n\nRaw content must be 32 bits. Base64 string must be 8 chars.",
                    "type": "string",
                    "format": "byte"
                },
                "mac": {
                    "description": "The MAC digest for this request, Base64 encodedn\n\nRaw content must be 256 bits. Base64 string must be 44 chars.",
                    "type": "string",
                    "format": "byte"
                }
            },
            "required": [
                "ebid",
                "time",
                "mac"
            ]
        },
...
    },
...
}